// packages/adhere-util-browsersniff/src/types.ts
export interface BrowserInfo {
  /** 浏览器名称 */
  browser: string;
  /** 设备类型 */
  device: string;
  /** 渲染引擎 */
  engine: string;
  /** 语言设置 */
  language: string;
  /** 操作系统 */
  os: string;
  /** 操作系统版本 */
  osVersion: string;
  /** 浏览器版本 */
  version: string;
}
// ...
export interface BrowserMatch {
  // 引擎
  Trident: boolean;
  Presto: boolean;
  WebKit: boolean;
  Gecko: boolean;
  KHTML: boolean;
  
  // 浏览器
  Safari: boolean;
  Chrome: boolean;
  IE: boolean;
  Edge: boolean;
  Firefox: boolean;
  'Firefox Focus': boolean;
  Chromium: boolean;
  Opera: boolean;
  Vivaldi: boolean;
  Yandex: boolean;
  Arora: boolean;
  Lunascape: boolean;
  QupZilla: boolean;
  'Coc Coc': boolean;
  Kindle: boolean;
  Iceweasel: boolean;
  Konqueror: boolean;
  Iceape: boolean;
  SeaMonkey: boolean;
  Epiphany: boolean;
  360: boolean;
  '360EE': boolean;
  '360SE': boolean;
  UC: boolean;
  QQBrowser: boolean;
  QQ: boolean;
  Baidu: boolean;
  Maxthon: boolean;
  Sogou: boolean;
  Liebao: boolean;
  '2345Explorer': boolean;
  '115Browser': boolean;
  TheWorld: boolean;
  XiaoMi: boolean;
  Quark: boolean;
  Qiyu: boolean;
  Wechat: boolean;
  WechatWork: boolean;
  Taobao: boolean;
  Alipay: boolean;
  Weibo: boolean;
  Douban: boolean;
  Suning: boolean;
  iQiYi: boolean;
  DingTalk: boolean;
  Huawei: boolean;
  Vivo: boolean;
  
  // 系统或平台
  Windows: boolean;
  Linux: boolean;
  'Mac OS': boolean;
  Android: boolean;
  HarmonyOS: boolean;
  Ubuntu: boolean;
  FreeBSD: boolean;
  Debian: boolean;
  'Windows Phone': boolean;
  BlackBerry: boolean;
  MeeGo: boolean;
  Symbian: boolean;
  iOS: boolean;
  'Chrome OS': boolean;
  WebOS: boolean;
  
  // 设备
  Mobile: boolean;
  Tablet: boolean;
}